+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added email or phone in Family Members list entry [#549](https://github.com/rokwire/safer-illinois-app/issues/549).
 
 ## [2.9.16] - 2021-03-02
 ### Fixed